% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binomialbcp.R
\name{binomialbcp.control}
\alias{binomialbcp.control}
\title{BB-BCP MCMC settings}
\usage{
binomialbcp.control(mcmc.iterations = 100, mcmc.burnin = 50,
  mcmc.thin = 5, verbose = 0)
}
\arguments{
\item{mcmc.iterations}{number of iterations}

\item{mcmc.burnin}{burnin samples}

\item{mcmc.thin}{thining interval}

\item{verbose}{if zero, no debug messages; if 1, once per mcmc step; if >1, log each Gibbs sample}
}
\description{
This function makes autocomplete work for \code{binomialbcp}.
}

