% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binomialbcp-package.R, R/binomialbcp.R
\docType{package}
\name{binomialbcp}
\alias{binomialbcp}
\alias{binomialbcp-package}
\title{Bayesian Changepoint for Binomial Data}
\usage{
binomialbcp(x, n, prior = list(a = 1, b = 1),
  control = binomialbcp.control())
}
\arguments{
\item{x}{the count of successes}

\item{n}{the size of each trial}

\item{prior}{a list of shape parameters for the beta prior}

\item{control}{a binomialBCP.control object}
}
\value{
a list containing the posterior samples of changepoints and expected values of proportions
}
\description{
This package provides a binomial variant of bayesian changepoint analysis.

Fit a beta-binomial bayesian changepoint model
}
\examples{
data(kansas)
 binomialbcp(kansas$Democrat, kansas$TotalVote)
}
\author{
Neal Fultz \email{njf@zestfinance.com},
}

