% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binomialbcp-package.R
\docType{data}
\name{kansas}
\alias{kansas}
\title{Kansas Presidential Election Data}
\format{A data frame with 38 obs of 6 variables:
\describe{
  \item{Year}{year of election}
  \item{TotalVote}{Number of Votes in 1000s}
  \item{Democrat}{}
  \item{Republican}{}
  \item{Independent}{}
  \item{Other}{}
}}
\source{
Dave Leip's Atlas of US Presidential Elections \url{http://uselectionatlas.org/RESULTS/compare.php?year=2012&fips=20&f=1&off=0&elect=0&type=state}
}
\description{
A data set of presidential election in the state of Kansas from 1864 to 2012. Votes are in 1000s.
}

