% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binomialbcp.R
\name{plot.binomialbcp}
\alias{plot.binomialbcp}
\title{Plot a Binomial BCP sample}
\usage{
\method{plot}{binomialbcp}(x, ...)
}
\arguments{
\item{x}{a binomialbcp sample}

\item{...}{passed on to plot}
}
\description{
Plot a Binomial BCP sample
}

