% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binomialbcp.R
\name{summary.binomialbcp}
\alias{summary.binomialbcp}
\title{Summarize a Binomial BCP sample}
\usage{
\method{summary}{binomialbcp}(object, ...)
}
\arguments{
\item{object}{a binomialbcp sample}

\item{...}{(unused)}
}
\value{
a dataframe containing the probability of a change point and estimated proportion for each time point.
}
\description{
Average over the MCMC samples to estimate changepoints and proportions.
}

